+++ server/src/navigation/assset.navigation-strategy.ts
@@ -1,23 +1,12 @@
 import { join } from 'path';
 import { AbstractNode } from '../parser/ast';
-import {
-    CosmoteerFile,
-    CosmoteerWorkspaceService,
-} from '../workspace/cosmoteer-workspace.service';
-import {
-    extractSubstrings,
-    filePathToDirectoryPath,
-    NavigationStrategy,
-} from './navigation-strategy';
+import { CosmoteerFile, CosmoteerWorkspaceService } from '../workspace/cosmoteer-workspace.service';
+import { extractSubstrings, filePathToDirectoryPath, NavigationStrategy } from './navigation-strategy';
 import { access, constants, readdir, realpath } from 'fs/promises';
 import { globalSettings } from '../server';
 
 export class AssetNavigationStrategy extends NavigationStrategy<boolean> {
-    async navigate(
-        path: string,
-        _startNode: AbstractNode,
-        currentLocation: string
-    ): Promise<boolean> {
+    async navigate(path: string, _startNode: AbstractNode, currentLocation: string): Promise<boolean> {
         if (path.startsWith('./Data')) {
             return this.navigateTroughCosmoteerFiles(path) ? true : false;
         } else {
@@ -25,25 +14,14 @@ export class AssetNavigationStrategy extends NavigationStrategy<boolean> {
         }
     }
     navigateTroughCosmoteerFiles(path: string): CosmoteerFile | null {
-        return CosmoteerWorkspaceService.instance.findFile(
-            extractSubstrings(path).slice(2)
-        ) as CosmoteerFile | null;
+        return CosmoteerWorkspaceService.instance.findFile(extractSubstrings(path).slice(2)) as CosmoteerFile | null;
     }
 
-    async navigateTroughOwnFiles(
-        path: string,
-        currentLocation: string
-    ): Promise<boolean> {
-        return await this.navigateRulesByCurrentLocation(
-            extractSubstrings(path),
-            currentLocation
-        );
+    async navigateTroughOwnFiles(path: string, currentLocation: string): Promise<boolean> {
+        return await this.navigateRulesByCurrentLocation(extractSubstrings(path), currentLocation);
     }
 
-    navigateRulesByCurrentLocation = async (
-        pathes: string[],
-        currentLocation: string
-    ) => {
+    navigateRulesByCurrentLocation = async (pathes: string[], currentLocation: string) => {
         try {
             const cleanedPath = join(currentLocation, '..', ...pathes);
 
@@ -61,22 +39,13 @@ export class AssetNavigationStrategy extends NavigationStrategy<boolean> {
     /**
      * Searches for a file in the directory incasesensitive which {access} will fail most of the times.
      */
-    private searchInDirForFile = async (
-        pathes: string[],
-        currentLocation: string
-    ): Promise<boolean> => {
+    private searchInDirForFile = async (pathes: string[], currentLocation: string): Promise<boolean> => {
         try {
             const realPath = await realpath(
-                join(
-                    filePathToDirectoryPath(currentLocation),
-                    ...pathes.slice(0, pathes.length - 1)
-                )
+                join(filePathToDirectoryPath(currentLocation), ...pathes.slice(0, pathes.length - 1))
             );
             const dir = await readdir(realPath);
-            const file = dir.find(
-                (f) =>
-                    f.toLowerCase() === pathes[pathes.length - 1].toLowerCase()
-            );
+            const file = dir.find((f) => f.toLowerCase() === pathes[pathes.length - 1].toLowerCase());
             return file ? true : false;
         } catch (error) {
             if (globalSettings.trace.server !== 'off') {