+++ server/src/navigation/navigation-strategy.ts
@@ -3,11 +3,7 @@ import { AbstractNode } from '../parser/ast';
 import { join } from 'path';
 
 export abstract class NavigationStrategy<T> {
-    abstract navigate(
-        path: string,
-        startNode: AbstractNode,
-        currentLocation: string
-    ): Promise<T>;
+    abstract navigate(path: string, startNode: AbstractNode, currentLocation: string): Promise<T>;
 }
 
 export const filePathToDirectoryPath = (path: string) => {
@@ -21,12 +17,7 @@ export const filePathToDirectoryPath = (path: string) => {
         return cleanedPath.substring(0, cleanedPath.lastIndexOf('/') + 1);
     }
     if (path.endsWith('.rules')) {
-        return path.substring(
-            0,
-            (path.includes('/')
-                ? path.lastIndexOf('/')
-                : path.lastIndexOf('\\') - 1) + 1
-        );
+        return path.substring(0, (path.includes('/') ? path.lastIndexOf('/') : path.lastIndexOf('\\') - 1) + 1);
     }
     return path;
 };