+++ server/src/utils/ast.utils.ts
@@ -1,19 +1,11 @@
 import { Position } from 'vscode-languageserver';
-import {
-    AbstractNode,
-    AbstractNodeDocument,
-    isArrayNode,
-    isAssignmentNode,
-    isObjectNode,
-} from '../parser/ast';
+import { AbstractNode, AbstractNodeDocument, isArrayNode, isAssignmentNode, isObjectNode } from '../parser/ast';
 import { FileWithPath } from '../workspace/cosmoteer-workspace.service';
 import { readFile } from 'fs/promises';
 import { lexer } from '../lexer/lexer';
 import { parser } from '../parser/parser';
 
-export const parseFile = async (
-    file: FileWithPath
-): Promise<AbstractNodeDocument> => {
+export const parseFile = async (file: FileWithPath): Promise<AbstractNodeDocument> => {
     const data = await readFile(file.path, { encoding: 'utf-8' });
     const document = parser(lexer(data), file.path).value;
     return document;
@@ -25,10 +17,7 @@ export const parseFilePath = async (path: string) => {
     return document;
 };
 
-export const findNodeAtPosition = (
-    document: AbstractNodeDocument,
-    position: Position
-) => {
+export const findNodeAtPosition = (document: AbstractNodeDocument, position: Position) => {
     for (const node of document.elements) {
         const foundNode = findNodeAtPositionRecursive(node, position);
         if (foundNode) {
@@ -37,18 +26,12 @@ export const findNodeAtPosition = (
     }
 };
 
-const findNodeAtPositionRecursive = (
-    node: AbstractNode,
-    position: Position
-): AbstractNode | undefined => {
+const findNodeAtPositionRecursive = (node: AbstractNode, position: Position): AbstractNode | undefined => {
     if (isObjectNode(node)) {
         for (const property of node.elements) {
             if (node.inheritance) {
                 for (const inheritance of node.inheritance) {
-                    const foundNode = findNodeAtPositionRecursive(
-                        inheritance,
-                        position
-                    );
+                    const foundNode = findNodeAtPositionRecursive(inheritance, position);
                     if (foundNode) {
                         return foundNode;
                     }
@@ -62,10 +45,7 @@ const findNodeAtPositionRecursive = (
     } else if (isArrayNode(node)) {
         if (node.inheritance) {
             for (const inheritance of node.inheritance) {
-                const foundNode = findNodeAtPositionRecursive(
-                    inheritance,
-                    position
-                );
+                const foundNode = findNodeAtPositionRecursive(inheritance, position);
                 if (foundNode) {
                     return foundNode;
                 }
@@ -105,4 +85,3 @@ export const getStartOfAstNode = (node: AbstractNode): AbstractNodeDocument => {
     }
     return node as AbstractNodeDocument;
 };
-