+++ server/src/utils/reference.utils.ts
@@ -1,12 +1,10 @@
-export const startsWithAmpersandAndLetter = (value: string) =>
-    /^&[A-Za-z_]/.test(value);
+export const startsWithAmpersandAndLetter = (value: string) => /^&[A-Za-z_]/.test(value);
 
 export const isValidReference = (value: string) => {
     if (value.startsWith('&')) {
         const valueWithoutAmpersand = value.substring(1);
         if (
-            (valueWithoutAmpersand.startsWith('<') &&
-                valueWithoutAmpersand.includes('.rules>')) ||
+            (valueWithoutAmpersand.startsWith('<') && valueWithoutAmpersand.includes('.rules>')) ||
             valueWithoutAmpersand.startsWith('..') ||
             valueWithoutAmpersand.startsWith('~') ||
             valueWithoutAmpersand.startsWith('/') ||