+++ server/src/validation/validator.functioncall.ts
@@ -9,29 +9,17 @@ export const ValidationForFunctionCall: Validation<FunctionCallNode> = {
             if (arg.type === 'Value' && arg.valueType.type === 'Reference') {
                 if (!(arg.valueType.value as string).startsWith('&')) {
                     return {
-                        message: l10n.t(
-                            'Reference in function calls need to start with an ampersand'
-                        ),
+                        message: l10n.t('Reference in function calls need to start with an ampersand'),
                         node: arg,
                     };
                 }
-                if (
-                    !arg.parenthesized &&
-                    arg.delimiter === undefined &&
-                    node.arguments.length > 1
-                ) {
+                if (!arg.parenthesized && arg.delimiter === undefined && node.arguments.length > 1) {
                     return {
-                        message: l10n.t(
-                            'Reference in function calls need to be parenthesized'
-                        ),
+                        message: l10n.t('Reference in function calls need to be parenthesized'),
                         node: arg,
                     };
                 }
-            } else if (
-                arg.type === 'Value' &&
-                arg.valueType.type !== 'Reference' &&
-                arg.valueType.type !== 'Number'
-            ) {
+            } else if (arg.type === 'Value' && arg.valueType.type !== 'Reference' && arg.valueType.type !== 'Number') {
                 return {
                     message: l10n.t(
                         'Invalid argument type, expected Reference(&) or Number. Got {0}',