+++ server/src/validation/validator.math.ts
@@ -1,8 +1,4 @@
-import {
-    AbstractNode,
-    isExpressionNode,
-    MathExpressionNode,
-} from '../parser/ast';
+import { AbstractNode, isExpressionNode, MathExpressionNode } from '../parser/ast';
 import { Validation } from './validator';
 import * as l10n from '@vscode/l10n';
 
@@ -12,36 +8,21 @@ export const ValidationForMath: Validation<MathExpressionNode> = {
         let lastNode: AbstractNode | null = null;
         for (const child of node.elements) {
             // Check for double expressions
-            if (
-                lastNode &&
-                isExpressionNode(lastNode) &&
-                isExpressionNode(child)
-            ) {
+            if (lastNode && isExpressionNode(lastNode) && isExpressionNode(child)) {
                 return {
-                    message: l10n.t(
-                        'Between two expressions should be a Value'
-                    ),
+                    message: l10n.t('Between two expressions should be a Value'),
                     node: child,
                 };
             }
             // Check for last element in MathExpression
-            if (
-                isExpressionNode(child) &&
-                node.elements[node.elements.length - 1] === child
-            ) {
+            if (isExpressionNode(child) && node.elements[node.elements.length - 1] === child) {
                 return {
-                    message: l10n.t(
-                        'Last element in MathExpression should be a Value'
-                    ),
+                    message: l10n.t('Last element in MathExpression should be a Value'),
                     node: child,
                 };
             }
             // Check for Value type
-            if (
-                child.type === 'Value' &&
-                child.valueType.type !== 'Number' &&
-                child.valueType.type !== 'Reference'
-            ) {
+            if (child.type === 'Value' && child.valueType.type !== 'Number' && child.valueType.type !== 'Reference') {
                 return {
                     message: l10n.t(
                         'Invalid argument type, expected Number or Reference. Got {0}',